+++ packages/terra-popup/src/_PopupContent.jsx
@@ -51,7 +51,7 @@ const propTypes = {
    */
   contentWidthMax: PropTypes.number,
   /**
-   * Should the popup content have tab focus. Set this is your content doesn't contain any focusable elements.
+   * Should the popup content have tab focus. Set this to `true` if your content doesn't contain any focusable elements.
    */
   isFocusedDisabled: PropTypes.bool,
   /**