+++ packages/terra-hookshot/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Improved focus management on iOS for hookshot containers which contain focusable elements
 
 4.12.0 - (September 25, 2018)
 ------------------